Add platform checker to handle dart:io incompatibility with Flutter web
This commit introduces a platform checker that uses conditional imports to determine whether the platform is Android or not. This is done to handle the incompatibility of dart:io with Flutter web, as dart:io is not available on the web. The platform checker defaults to false (not Android and not Mobile) if dart:io is not available.